+++ app/(gcforms)/[locale]/(form filler)/id/[...props]/lib/exceptions.ts
@@ -1,0 +1,8 @@
+export class MissingFormIdError extends Error {}
+export class MissingFormDataError extends Error {}
+export class FormNotFoundError extends Error {}
+export class FormIsClosedError extends Error {}
+export class FileSizeError extends Error {}
+export class FileTypeError extends Error {}
+export class FileObjectInvalid extends Error {}
+export class SubmissionApiError extends Error {}
